3.4.1

+++ proxy-vpn-blocker/includes/class-proxy-vpn-blocker.php
@@ -114,7 +114,7 @@ class Proxy_VPN_Blocker {
 	 * @since   1.0
 	 * @return  void
 	 */
-	public function __construct( $file = '', $version = '3.4.0' ) {
+	public function __construct( $file = '', $version = '3.4.1' ) {
 		$this->_version = $version;
 		$this->_token   = 'proxy_vpn_blocker';
 
@@ -364,7 +364,7 @@ class Proxy_VPN_Blocker {
 	 * @see proxy_vpn_blocker()
 	 * @return Main proxy_vpn_blocker instance
 	 */
-	public static function instance( $file = '', $version = '3.4.0' ) {
+	public static function instance( $file = '', $version = '3.4.1' ) {
 
 		if ( is_null( self::$_instance ) ) {
 			self::$_instance = new self( $file, $version );
@@ -424,6 +424,11 @@ class Proxy_VPN_Blocker {
 	 * @return void
 	 */
 	public function redirect_after_activation() {
+		// Don't redirect during AJAX requests.
+		if ( wp_doing_ajax() ) {
+			return;
+		}
+
 		if ( get_option( $this->_token . '_pvb_activation_redirect', false ) ) {
 			delete_option( $this->_token . '_pvb_activation_redirect' );
